Better Header with placeholder text from TwUI
