+++ public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "short_name": "React App",
-  "name": "Create React App Sample",
+  "short_name": "Gurkult",
+  "name": "Gurkult - A Kult for open source",
   "icons": [
     {
       "src": "favicon.ico",
