+++ public/robots.txt
@@ -1,3 +1,8 @@
-# https://www.robotstxt.org/robotstxt.html
-User-agent: *
+User-agent: Xithrius
 Disallow:
+
+User-agent: Googlebot
+Allow: /
+
+User-agent: *
+Disallow: /
