+++ src/components/hero.tsx
@@ -1,6 +1,19 @@
-import Navbar from './navbar'
+/* This example requires Tailwind CSS v2.0+ */
+import { Fragment } from 'react'
+import { Popover, Transition } from '@headlessui/react'
+import { MenuIcon, XIcon } from '@heroicons/react/outline'
+import { serverInvite, github, shop } from '../constants';
+import gurks from '../assets/gurks.png'
+import crown from '../assets/crown.png'
 
-export default function Hero() {
+
+const navigation = [
+  { name: "Discord", href: `${serverInvite}`, current: false },
+  { name: "Contribute", href: `${github}`, current: false },
+  { name: "Store", href: `${shop}`, current: false },
+]
+
+export default function Example() {
   return (
     <div className="relative bg-white overflow-hidden">
       <div className="max-w-7xl mx-auto">
@@ -14,21 +27,116 @@ export default function Hero() {
           >
             <polygon points="50,0 100,0 50,100 0,100" />
           </svg>
-          <Navbar />
+
+          <Popover>
+            {({ open }) => (
+              <>
+                <div className="relative pt-6 px-4 sm:px-6 lg:px-8">
+                  <nav
+                    className="relative flex items-center justify-between sm:h-10 lg:justify-start"
+                    aria-label="Global"
+                  >
+                    <div className="flex items-center flex-grow flex-shrink-0 lg:flex-grow-0">
+                      <div className="flex items-center justify-between w-full md:w-auto">
+                        <a href="/">
+                          <span className="sr-only">Workflow</span>
+                          <img
+                            className="h-8 w-auto sm:h-10"
+                            src={crown}
+                            alt="Gurkan Crown"
+                          />
+                        </a>
+                        <div className="-mr-2 flex items-center md:hidden">
+                          <Popover.Button className="bg-white rounded-md p-2 inline-flex items-center justify-center text-gray-400 hover:text-gray-500 hover:bg-gray-100 focus:outline-none focus:ring-2 focus:ring-inset focus:ring-indigo-500">
+                            <span className="sr-only">Open main menu</span>
+                            <MenuIcon className="h-6 w-6" aria-hidden="true" />
+                          </Popover.Button>
+                        </div>
+                      </div>
+                    </div>
+                    <div className="hidden md:block md:ml-10 md:pr-4 md:space-x-8">
+                      {navigation.map((item) => (
+                        <a key={item.name} href={item.href} className="font-medium text-gray-500 hover:text-gray-900">
+                          {item.name}
+                        </a>
+                      ))}
+                      <a href={serverInvite} className="font-medium text-green-600 hover:text-green-500">
+                        Join Us
+                      </a>
+                    </div>
+                  </nav>
+                </div>
+
+                <Transition
+                  show={open}
+                  as={Fragment}
+                  enter="duration-150 ease-out"
+                  enterFrom="opacity-0 scale-95"
+                  enterTo="opacity-100 scale-100"
+                  leave="duration-100 ease-in"
+                  leaveFrom="opacity-100 scale-100"
+                  leaveTo="opacity-0 scale-95"
+                >
+                  <Popover.Panel
+                    focus
+                    static
+                    className="absolute top-0 inset-x-0 p-2 transition transform origin-top-right md:hidden"
+                  >
+                    <div className="rounded-lg shadow-md bg-white ring-1 ring-black ring-opacity-5 overflow-hidden">
+                      <div className="px-5 pt-4 flex items-center justify-between">
+                        <div>
+                          <img
+                            className="h-8 w-auto"
+                            src={crown}
+                            alt="Gurkan Crown"
+                          />
+                        </div>
+                        <div className="-mr-2">
+                          <Popover.Button className="bg-white rounded-md p-2 inline-flex items-center justify-center text-gray-400 hover:text-gray-500 hover:bg-gray-100 focus:outline-none focus:ring-2 focus:ring-inset focus:ring-indigo-500">
+                            <span className="sr-only">Close main menu</span>
+                            <XIcon className="h-6 w-6" aria-hidden="true" />
+                          </Popover.Button>
+                        </div>
+                      </div>
+                      <div className="px-2 pt-2 pb-3 space-y-1">
+                        {navigation.map((item) => (
+                          <a
+                            key={item.name}
+                            href={item.href}
+                            className="block px-3 py-2 rounded-md text-base font-medium text-gray-700 hover:text-gray-900 hover:bg-gray-50"
+                          >
+                            {item.name}
+                          </a>
+                        ))}
+                      </div>
+                      <a
+                        href={serverInvite}
+                        className="block w-full px-5 py-3 text-center font-medium text-green-600 bg-gray-50 hover:bg-gray-100"
+                      >
+                        Join Us
+                      </a>
+                    </div>
+                  </Popover.Panel>
+                </Transition>
+              </>
+            )}
+          </Popover>
+
           <main className="mt-10 mx-auto max-w-7xl px-4 sm:mt-12 sm:px-6 md:mt-16 lg:mt-20 lg:px-8 xl:mt-28">
             <div className="sm:text-center lg:text-left">
               <h1 className="text-4xl tracking-tight font-extrabold text-gray-900 sm:text-5xl md:text-6xl">
-                <span className="block xl:inline">Lorem ipsum dolor<br /></span>
-                <span className="block text-green-400 xl:inline">sit amet consectetur</span>
+                <span className="block xl:inline">Data to enrich your</span>{' '}
+                <span className="block text-green-500 xl:inline">online business</span>
               </h1>
               <p className="mt-3 text-base text-gray-500 sm:mt-5 sm:text-lg sm:max-w-xl sm:mx-auto md:mt-5 md:text-xl lg:mx-0">
-                adipiscing elit. Integer elementum nec turpis ut lobortis. Nulla sed ultrices enim. Nam auctor leo quis risus pretium, non molestie tellus tempus.
+                Anim aute id magna aliqua ad ad non deserunt sunt. Qui irure qui lorem cupidatat commodo. Elit sunt amet
+                fugiat veniam occaecat fugiat aliqua.
               </p>
               <div className="mt-5 sm:mt-8 sm:flex sm:justify-center lg:justify-start">
                 <div className="rounded-md shadow">
                   <a
                     href="/"
-                    className="w-full flex items-center justify-center px-8 py-3 border border-transparent text-base font-medium rounded-md text-white bg-green-400 hover:bg-green-500 md:py-4 md:text-lg md:px-10"
+                    className="w-full flex items-center justify-center px-8 py-3 border border-transparent text-base font-medium rounded-md text-white bg-green-500 hover:bg-green-400 md:py-4 md:text-lg md:px-10"
                   >
                     Get started
                   </a>
@@ -36,7 +144,7 @@ export default function Hero() {
                 <div className="mt-3 sm:mt-0 sm:ml-3">
                   <a
                     href="/"
-                    className="w-full flex items-center justify-center px-8 py-3 border border-transparent text-base font-medium rounded-md text-indigo-700 bg-indigo-100 hover:bg-indigo-200 md:py-4 md:text-lg md:px-10"
+                    className="w-full flex items-center justify-center px-8 py-3 border border-transparent text-base font-medium rounded-md text-green-700 bg-green-100 hover:bg-green-200 md:py-4 md:text-lg md:px-10"
                   >
                     Live demo
                   </a>
@@ -49,8 +157,8 @@ export default function Hero() {
       <div className="lg:absolute lg:inset-y-0 lg:right-0 lg:w-1/2">
         <img
           className="h-56 w-full object-cover sm:h-72 md:h-96 lg:w-full lg:h-full"
-          src="https://images.unsplash.com/photo-1551434678-e076c223a692?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=2850&q=80"
-          alt=""
+          src={gurks}
+          alt="Gurks"
         />
       </div>
     </div>
