+++ src/pages/App.tsx
@@ -1,5 +1,4 @@
 import React from "react";
-import Navbar from "../components/navbar"
 import Hero from "../components/hero";
 import gurkultLogo from "../assets/logo.png";
